refactor: 상품 관리 분리
